Add basic scaffolds for business index listing, create, read, and update
This does not yet include images. It is also unstyled and does not require
login. However, it sets up the relationship to profile, allows basic CRUD
actions, and generally can be expanded upon easily. This is met to just be a
first step. From here I plan on adding the images capabilities, getting some
basic styling using the CSS that already exists, and continuing to add the links
to this as needed per the tickets elsewhere on the site.

Ticket:
[1](https://app.clubhouse.io/peakbuildersnetwork/story/264/create-page-business-directory)
- Done except for image

Up next:
- Get image added
- [Create linking properly upon signup](https://app.clubhouse.io/peakbuildersnetwork/story/83/create-links-on-member-profile-page-create-manage-business-listing)
- [Create page for listing that has more data and is set up per this flow](https://app.clubhouse.io/peakbuildersnetwork/story/262/create-page-list-a-business)